Make it play better with the help printing; there seems to be an error with how timestamp flags are handled.
